# openapi.yaml
openapi: 3.1.0
info:
  title: Connectors Engine API
  version: "0.1.0"
  description: |
    Rails engine exposing n8n's connections / credentials layer as a host-mountable API.
    Every endpoint is namespaced under the engine's mount path (default `/connectors`).

    **Auth model.** The engine itself doesn't authenticate requests — the host wires
    `Connectors.configuration.current_owner_resolver` to extract the requester from the
    controller (session, JWT, API key — host's choice). Endpoints that mutate
    credentials require an `Owner` to be resolvable; those that read enforce
    role-based visibility via `principal_resolver` (Phase 9).

    **Source of truth.** This document is hand-maintained alongside the engine.
    Cross-check against `connectors/config/routes.rb` and the controller files —
    they are the implementation.

    **Mount path.** The host picks where the engine lives. Reference deployment
    (flow-api) uses `/api/v1/connectors`. The engine introspects its actual
    mount via `Connectors::Engine.mount_path`, so OAuth callback URLs and
    webhook subscribe URLs in responses always reflect reality.
  contact:
    name: Connectors framework
  license:
    name: MIT
servers:
  - url: "{host}{mount}"
    description: Engine mount in the host app
    variables:
      host:
        default: "http://localhost:3000"
        description: Scheme + host of the host application
      mount:
        default: "/api/v1/connectors"
...
paths:
...
  /credentials:
    get:
      tags: [Credentials]
      summary: List credentials visible to the requester (owned + shared)
      operationId: listCredentials
      parameters:
        - { name: type,         in: query, schema: { type: string } }
        - { name: include_data, in: query, description: "Owner-only decrypted data; MCP returns public configuration for all visible connections", schema: { type: boolean, default: false } }
      responses:
        "200":
          description: Visible credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  credentials:
                    type: array
                    items: { $ref: "#/components/schemas/Credential" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Credentials]
      summary: Create a credential (paste-the-key flow)
      operationId: createCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:         { type: string, description: "connector key" }
                name:         { type: string }
                data:         { type: object, additionalProperties: true, description: "credential values" }
                external_ref: { type: string, description: "vault reference (Phase 10)" }
                is_managed:   { type: boolean, default: false }
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Credential" } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /credentials/{id}/actions/{name}:
    parameters:
      - { name: id,   in: path, required: true, schema: { type: string, format: uuid } }
      - { name: name, in: path, required: true, schema: { type: string  } }
    post:
      tags: [Actions]
      summary: Invoke a connector action with this credential
      description: |
        Validates `data` against the action's param schema, then runs the
        action's execute block in the connector instance's context. Returns
        a normalized envelope — `status: "ok"` with `data`, or `status: "error"`
        with a typed `error` object (`invalid_params`, `authentication_failed`,
        `forbidden`, `rate_limited`, `api_error`).
      operationId: invokeCredentialAction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                data: { type: object, additionalProperties: true, description: "action input params" }
      responses:
        "200":
          description: Action succeeded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403":
          description: Provider permission denied
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: Invalid params
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
        "429":
          description: Provider rate-limited the request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
        "502":
          description: Provider returned an error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
...
components:
  schemas:
...
    ActionResult:
      type: object
      description: "envelope returned by every action invocation — same shape on success and failure"
      properties:
        status: { type: string, enum: [ok, error] }
        action: { type: string, description: "action key" }
        data:   { type: object, additionalProperties: true, description: "present when status=ok" }
        error:
          type: object
          description: "present when status=error"
          properties:
            type:
              type: string
              enum: [invalid_params, authentication_failed, forbidden, rate_limited, api_error]
            message:    { type: string }
            status:     { type: integer, nullable: true, description: "provider HTTP status" }
            missing:    { type: array, items: { type: string }, description: "invalid_params: required keys not supplied" }
            unknown:    { type: array, items: { type: string }, description: "invalid_params: keys not on the schema" }
            retry_after: { type: integer, nullable: true, description: "rate_limited: seconds until next retry" }
            body: { type: object, additionalProperties: true, nullable: true, description: "api_error: raw provider response" }
...
    Credential:
      type: object
      properties:
        id:                       { type: string, format: uuid }
        type:                     { type: string, description: "connector key (n8n vocabulary)" }
        connector_key:            { type: string, deprecated: true }
        name:                     { type: string, description: "display name" }
        external_account_id:      { type: string, nullable: true }
        status:                   { type: string, enum: [active, expired, revoked, errored] }
        expires_at:               { type: string, format: date-time, nullable: true }
        last_used_at:             { type: string, format: date-time, nullable: true }
        created_at:               { type: string, format: date-time }
        updated_at:               { type: string, format: date-time }
        scopes:                   { type: array, items: { type: string }, nullable: true }
        is_managed:               { type: boolean }
        external_ref:             { type: string, nullable: true }
        __overwritten_properties: { type: array, items: { type: string } }
        data:
          type: object
          additionalProperties: true
          description: "With include_data=true, decrypted data for the owner role only; shared viewers/editors omit data. MCP returns only public configuration for all visible connections."
...
    Error:
      type: object
      properties:
        error: { type: string }
  responses:
...
    Unauthorized:
      description: Unauthorized (missing owner, role gate, or signature)
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource or connector type not found
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...